--- stops/apps/frontend/src/components/Stop/GenericContainer/index.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import MapContainer from '../MapContainer';
-import styles from '../styles.module.css';
-
-/* * */
-
-export default function GenericContainer() {
-	//
-
-	//
-	// A. Render components
-
-	return (
-		<div className={styles.container}>
-			<MapContainer generic={true} />
-		</div>
-	);
-}
