+++ stops/apps/frontend/src/components/StopContainer/MapContainer/Header/List/Item/index.tsx
