+++ stops/apps/frontend/src/components/StopContainer/MapContainer/Header/PatternsModal/index.tsx
