+++ stops/apps/frontend/src/components/StopContainer/StopAffectation/index.tsx
