+++ stops/apps/frontend/src/components/StopContainer/StopDetails/index.tsx
