+++ stops/apps/frontend/src/components/StopContainer/StopEquipments/index.tsx
