+++ stops/apps/frontend/src/components/StopContainer/StopEquipments/styles.module.css
