+++ stops/apps/frontend/src/components/StopContainer/StopMedia/index.tsx
