+++ stops/apps/frontend/src/components/StopContainer/StopObservations/styles.module.css
