+++ stops/apps/frontend/src/components/StopContainer/styles.module.css
