+++ stops/apps/frontend/tsconfig.json
@@ -19,7 +19,7 @@
 		"**/*.ts",
 		"**/*.tsx",
 		".next/types/**/*.ts"
-, "src/components/Stop/MapContainer/Header", "src/components/Stop/MapContainer/Mapper"	],
+, "src/components/StopContainer/MapContainer/Header", "src/components/StopContainer/MapContainer/Mapper"	],
 	"exclude": [
 		"node_modules"
 	]
